# docs/swagger.yaml
definitions:
...
  types.Command:
    description: an instruction to be executed by a pipeline
    properties:
      description:
        description: the command description
        example: scan files in specified path
        type: string
      error_regex:
        description: a regex used to determine if the command execution has errored
        example: '.*Infected files: [^0].*'
        type: string
      name:
        description: a unique name for the command
        example: clamscan
        type: string
      stop_on_error:
        description: determines if the process should stop on a command execution
          error
        example: true
        type: boolean
      value:
        description: the value of the command
        example: freshclam && clamscan -r ${path}
        type: string
    type: object
...
  types.Key:
    description: a digital key used to sign or verify signatures
    properties:
      description:
        description: a description of the intended use of the key
        type: string
      name:
        description: a unique identifier for the digital key
        type: string
      owner:
        description: the name of the entity owning the key
        type: string
      value:
        description: the actual content of the key
        type: string
    type: object
  types.Notification:
    properties:
      name:
        description: unique identifier for the notification
        type: string
      recipient:
        description: Recipient of the notification if type is email
        type: string
      template:
        description: Template to use for content of the notification
        type: string
      type:
        description: Type of the notification (e.g. email, snow, etc.)
        type: string
    type: object
  types.NotificationTemplate:
    properties:
      content:
        description: Content of the template
        type: string
      name:
        description: Name unique identifier for notification template
        type: string
      subject:
        description: Subject of the notification
        type: string
    type: object
...
info:
  contact:
    email: info@southwinds.io
    name: SouthWinds Tech Ltd
    url: https://www.southwinds.io/
  description: Securely transfer artisan packages and container images between repositories
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Artisan's Doorman
  version: 1.0.0
paths:
...
  /command:
    put:
      consumes:
      - application/yaml
      - ' application/json'
      description: creates or updates a command
      parameters:
      - description: the data for the command to persist
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/types.Command'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Creates or updates a command
      tags:
      - Commands
  /event/{service-id}/{bucket-name}/{folder-name}:
    post:
      description: Triggers the ingestion of a specification
      parameters:
      - description: a unique identifier for the bucket endpoint (e.g. x-minio-deployment-id
          for MinIO)
        in: path
        name: service-id
        required: true
        type: string
      - description: the name of the bucket that contains the uploaded files
        in: path
        name: bucket-name
        required: true
        type: string
      - description: the name of the folder within the bucket that contains the uploaded
          files
        in: path
        name: folder-name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Triggers the ingestion of an artisan spec artefacts
      tags:
      - Webhook
...
  /job:
    get:
      description: Gets a list of top job information logs
      parameters:
      - description: the number of top jobs to retrieve, if not specified returns
          the top 1 job
        in: query
        name: count
        type: integer
      produces:
      - application/json
      - ' application/yaml'
      - ' application/xml'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Gets top jobs
      tags:
      - Jobs
  /key:
    put:
      consumes:
      - application/yaml
      - ' application/json'
      description: |-
        creates or updates a cryptographic key used by either inbound or outbound routes to verify or sign
        packages respectively
      parameters:
      - description: the data for the key to persist
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/types.Key'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Creates or updates a cryptographic key
      tags:
      - Keys
  /notification:
    get:
      description: gets all notifications
      produces:
      - application/json
      - ' application/yaml'
      - ' application/xml'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Gets all notifications
      tags:
      - Notifications
    put:
      consumes:
      - application/yaml
      - ' application/json'
      description: creates or updates a notification
      parameters:
      - description: the data for the notification to persist
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/types.Notification'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Creates or updates a notification
      tags:
      - Notifications
  /notification-template:
    get:
      description: gets all notification templates
      produces:
      - application/json
      - ' application/yaml'
      - ' application/xml'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Gets all notification templates
      tags:
      - Notifications
    put:
      consumes:
      - application/yaml
      - ' application/json'
      description: creates or updates a notification template
      parameters:
      - description: the data for the notification template to persist
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/types.NotificationTemplate'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Creates or updates a notification template
      tags:
      - Notifications
...
  /pipe:
    get:
      description: gets all pipelines
      produces:
      - application/json
      - ' application/yaml'
      - ' application/xml'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Gets all pipelines
      tags:
      - Pipelines
    put:
      consumes:
      - application/yaml
      - ' application/json'
      description: creates or updates an inbound route
      parameters:
      - description: the data for the pipeline to persist
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/types.PipelineConf'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Creates or updates an inbound route
      tags:
      - Pipelines
  /pipe/{name}:
    get:
      description: gets a pipeline
      parameters:
      - description: the name of the pipeline to retrieve
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      - ' application/yaml'
      - ' application/xml'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Gets a pipeline
      tags:
      - Pipelines
  /route/in:
    put:
      consumes:
      - application/yaml
      - ' application/json'
      description: creates or updates an inbound route
      parameters:
      - description: the data for the inbound route to persist
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/types.InRoute'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Creates or updates an inbound route
      tags:
      - Routes
  /route/out:
    put:
      consumes:
      - application/yaml
      - ' application/json'
      description: creates or updates an inbound route
      parameters:
      - description: the data for the outbound route to persist
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/types.OutRoute'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Creates or updates an inbound route
      tags:
      - Routes
  /token:
    get:
      description: |-
        get authentication information for all webhook tokens
        an inbound route, returning required referrer URL and IP white list
        NOTE: this endpoint is called by the proxy to authenticate its webhook
      produces:
      - application/json
      - ' application/yaml'
      - ' application/xml'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get authentication information for all webhook tokens
      tags:
      - Webhook
...
swagger: "2.0"